Rename layer screen to render
